Don't export EditorDocumentManager as a Roslyn IWorkspaceService
