+++ src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/Documents/RunningDocumentTableEventSink.cs
@@ -1,22 +1,16 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-using System;
 using Microsoft.VisualStudio.Shell.Interop;
 
 namespace Microsoft.VisualStudio.Editor.Razor.Documents;
 
-internal class RunningDocumentTableEventSink : IVsRunningDocTableEvents3
+internal sealed class RunningDocumentTableEventSink : IVsRunningDocTableEvents3
 {
     private readonly VisualStudioEditorDocumentManager _documentManager;
 
     public RunningDocumentTableEventSink(VisualStudioEditorDocumentManager documentManager)
     {
-        if (documentManager is null)
-        {
-            throw new ArgumentNullException(nameof(documentManager));
-        }
-
         _documentManager = documentManager;
     }
