+++ src/Razor/test/Microsoft.VisualStudio.Editor.Razor.Test/Documents/EditorDocumentManagerListenerTest.cs
@@ -9,7 +9,6 @@ using Microsoft.AspNetCore.Razor.Test.Common.ProjectSystem;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Razor;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
-using Microsoft.CodeAnalysis.Razor.Workspaces;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.Threading;
 using Moq;
@@ -48,7 +47,7 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
         var opened = new EventHandler((o, args) => { });
         var closed = new EventHandler((o, args) => { });
 
-        var editorDocumentManger = new Mock<EditorDocumentManager>(MockBehavior.Strict);
+        var editorDocumentManger = new Mock<IEditorDocumentManager>(MockBehavior.Strict);
         var document = GetEditorDocument(documentManager: editorDocumentManger.Object);
         editorDocumentManger
             .Setup(e => e.TryGetDocument(It.IsAny<DocumentKey>(), out document))
@@ -57,10 +56,8 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
             .Setup(e => e.RemoveDocument(It.IsAny<EditorDocument>()))
             .Callback<EditorDocument>(doc => Assert.Same(document, doc));
 
-        var workspaceProvider = Mock.Of<IWorkspaceProvider>(MockBehavior.Strict);
-
         var listener = new EditorDocumentManagerListener(
-            workspaceProvider, Dispatcher, JoinableTaskContext, editorDocumentManger.Object, changedOnDisk, changedInEditor, opened, closed);
+            editorDocumentManger.Object, Dispatcher, JoinableTaskContext, changedOnDisk, changedInEditor, opened, closed);
 
         var projectFilePath = "/Path/to/project.csproj";
         var project = Mock.Of<IProjectSnapshot>(p => p.Key == TestProjectKey.Create("/Path/to/obj") && p.FilePath == projectFilePath, MockBehavior.Strict);
@@ -78,7 +75,7 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
         var opened = new EventHandler((o, args) => { });
         var closed = new EventHandler((o, args) => { });
 
-        var editorDocumentManger = new Mock<EditorDocumentManager>(MockBehavior.Strict);
+        var editorDocumentManger = new Mock<IEditorDocumentManager>(MockBehavior.Strict);
         var document = GetEditorDocument(documentManager: editorDocumentManger.Object);
         editorDocumentManger
             .Setup(e => e.TryGetDocument(It.IsAny<DocumentKey>(), out document))
@@ -87,10 +84,8 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
             .Setup(e => e.RemoveDocument(It.IsAny<EditorDocument>()))
             .Callback<EditorDocument>(doc => Assert.Same(document, doc));
 
-        var workspaceProvider = Mock.Of<IWorkspaceProvider>(MockBehavior.Strict);
-
         var listener = new EditorDocumentManagerListener(
-            workspaceProvider, Dispatcher, JoinableTaskContext, editorDocumentManger.Object, changedOnDisk, changedInEditor, opened, closed);
+            editorDocumentManger.Object, Dispatcher, JoinableTaskContext, changedOnDisk, changedInEditor, opened, closed);
 
         var projectFilePath = "/Path/to/project.csproj";
         var project = Mock.Of<IProjectSnapshot>(p =>
@@ -111,7 +106,7 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
         var opened = new EventHandler((o, args) => { });
         var closed = new EventHandler((o, args) => { });
 
-        var editorDocumentManger = new Mock<EditorDocumentManager>(MockBehavior.Strict);
+        var editorDocumentManger = new Mock<IEditorDocumentManager>(MockBehavior.Strict);
         editorDocumentManger
             .Setup(e => e.GetOrCreateDocument(It.IsAny<DocumentKey>(), It.IsAny<string>(), It.IsAny<ProjectKey>(), It.IsAny<EventHandler>(), It.IsAny<EventHandler>(), It.IsAny<EventHandler>(), It.IsAny<EventHandler>()))
             .Returns(GetEditorDocument())
@@ -123,10 +118,8 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
                 Assert.Same(closed, onClosed);
             });
 
-        var workspaceProvider = Mock.Of<IWorkspaceProvider>(MockBehavior.Strict);
-
         var listener = new EditorDocumentManagerListener(
-            workspaceProvider, Dispatcher, JoinableTaskContext, editorDocumentManger.Object, changedOnDisk, changedInEditor, opened, closed);
+            editorDocumentManger.Object, Dispatcher, JoinableTaskContext, changedOnDisk, changedInEditor, opened, closed);
 
         var projectFilePath = "/Path/to/project.csproj";
         var project = Mock.Of<IProjectSnapshot>(p => p.Key == TestProjectKey.Create("/Path/to/obj") && p.FilePath == projectFilePath, MockBehavior.Strict);
@@ -142,15 +135,13 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
         var called = false;
         var opened = new EventHandler((o, args) => called = true);
 
-        var editorDocumentManger = new Mock<EditorDocumentManager>(MockBehavior.Strict);
+        var editorDocumentManger = new Mock<IEditorDocumentManager>(MockBehavior.Strict);
         editorDocumentManger
             .Setup(e => e.GetOrCreateDocument(It.IsAny<DocumentKey>(), It.IsAny<string>(), It.IsAny<ProjectKey>(), It.IsAny<EventHandler>(), It.IsAny<EventHandler>(), It.IsAny<EventHandler>(), It.IsAny<EventHandler>()))
             .Returns(GetEditorDocument(isOpen: true));
 
-        var workspaceProvider = Mock.Of<IWorkspaceProvider>(MockBehavior.Strict);
-
         var listener = new EditorDocumentManagerListener(
-            workspaceProvider, Dispatcher, JoinableTaskContext, editorDocumentManger.Object, onChangedOnDisk: null, onChangedInEditor: null, onOpened: opened, onClosed: null);
+            editorDocumentManger.Object, Dispatcher, JoinableTaskContext, onChangedOnDisk: null, onChangedInEditor: null, onOpened: opened, onClosed: null);
 
         var projectFilePath = "/Path/to/project.csproj";
         var project = Mock.Of<IProjectSnapshot>(p => p.Key == TestProjectKey.Create("/Path/to/obj") && p.FilePath == projectFilePath, MockBehavior.Strict);
@@ -162,10 +153,10 @@ public class EditorDocumentManagerListenerTest : ProjectSnapshotManagerDispatche
         Assert.True(called);
     }
 
-    private EditorDocument GetEditorDocument(bool isOpen = false, EditorDocumentManager? documentManager = null)
+    private EditorDocument GetEditorDocument(bool isOpen = false, IEditorDocumentManager? documentManager = null)
     {
         var document = new EditorDocument(
-            documentManager ?? Mock.Of<EditorDocumentManager>(MockBehavior.Strict),
+            documentManager ?? Mock.Of<IEditorDocumentManager>(MockBehavior.Strict),
             Dispatcher,
             JoinableTaskContext,
             _projectFilePath,
